Okay, I've put all out php code so far into the integrated.php document.
It's ordered with Mohammed Maroufi's code (to control the valve) at the top, then Willy'b's (to select the valve) and then finally Abdu's (to view the database). This integrated.php may not work at the moment but it's okay. We can run through it in person when the weather is better :)  So far the main thing Is to figure out how to migrate WillyB's and Mohammed's code to work with Abdu's back end tables. But we'll sort that out in person.